Add support for a user's guide
Pass -g/--guide args with files/dirs on the command line or guide_files to the
compile_docs functions. Files must be in markdown and if a dir is passed the
files in it will be put in the documentation alphabetically, but only for files
ending in ".md"

Note that having two files with the same basename is a bad idea and will not
work (the HTML gets written to the same file).