Add detail type and self and to top links
The detail type makes it slightly easier to see what you're dealing with when
the "Function summary" header is out of sight. The # link makes it easier to
get a link to the function and the to top link will come in handy once the
sidebar goes away to get back to the main navigation section.