Load Jquery from CDN only when using a remote site
A javascript function "should_request_from_CDN" has been created to determine
whether to load from the CDN or from a local copy.
The rules are. if the page is served via http or https, and the host is
not a private network (127.x.x.x, or 192.168.x.x., or others), and
the hostname is contains a "."  (to avoid loading from localhost and others
user may have configured), only then, it will load from the CDN.

Range of private networks I am using: https://en.wikipedia.org/wiki/Private_network#Private_IPv4_address_spaces

I have used parts of this code, to determine if it's a private network
https://gist.github.com/dperini/729294

The other change is that now it requests from "//" and not from "https://"