.travis.yml: automatically download latest pandoc
travis-before-install.sh created.
it checks the latest stable relesase
based on this redirect url: https://github.com/jgm/pandoc/releases/latest
then, we parse the page, and determine look for the *-amd64.deb file.
So long as there is a file that matches this naming convention it will be retrieved,
and if may are provided, the firt one will be chosen.

If no file could be determined, then an error message is displayed.