fix(ExDoc.Retriever): fix invalid arguments for callbacks
When a callback is defined, an attempt is made to derive the argument
type from the name. If the typespec has more than one type option then
the argument name is given as "|"

An explicit check for this format is now in place, and the "arg#{i}"
case is used.

Fixes #476