Fix back button on Chrome, closes #593
There were two issues that caused the back button to break.

First, by setting the .main and .content to height: 100%
and overflow to hidden, the body of the page was no longer
the one effectively doing the scrolling, which broke Chrome.

Once that was fixed, the scrolling animation that happened by
clicking on the menu still broke the back button on Chrome
because Chrome saves the last page stage before the URL changes.
Since we would scroll to the #fragment before changing the URL,
Chrome will always redirect to the same #fragment if the back
button is used, given once again the impression it is broken.