Remove version name from epub formatter
Previously, the epub formatter defaulted to a filename in the following format:

> Phoenix-v1.3.0.epub

This commit changes it to:

> Phoenix.epub

When published to hexdocs, the version will already be specified in the url.
This means if the epub file is linked to in the docs, it will not be updated
each time there is a new version.