Rewrite html formatter tests so it avoid future problems with Travis
Tests will fail in Travis when a PR is on the works, and master is ahead
of the PR and new JS and CSS assets have been introduced to master.
You can see the errors here: https://travis-ci.org/elixir-lang/ex_doc/jobs/555347013#L422-L422
Travis fetches the PR and applies it to master.