(epub): Avoid duplicate IDs on Package Definition
Currently the ID generation for static files is based on the file name,
but the JS/CSS assets included by ExDoc have the same prefix, example:

* `epub-<hash>.js`
* `epub-<hash>.css`

This will produce duplicate IDs in the Package Document Definition,
which is an error.

Now we also take into account the extension for the static files.