Hide sidebar-over-content on click/tap outside it
On a small display, clicking anchor links in the sidebar (e.g. a
function while you're on that module page already) would navigate to
the anchor but not close the sidebar, requiring you to first click
the anchor and then the sidebar toggle.

José feels we shouldn't close the sidebar on the anchor click, because
that would be frustrating in case of misclicks, so instead we make the
sidebar easier to dismiss.

This is only in effect when the sidebar shows on top of the content, and
not on wide displays (where you might click the content just to
highlight some text, and wouldn't expect the sidebar to close).

Refs #1194.