Add translate="no" tags throughout the project
We skip translation only on names of modules, types,
functions, and code snippets. For example, the
subsection inside a module page are still translated.
The rationale is that terms like "function" and "macro"
appear as is in the text, without code tags, therefore
they are translated, so they same should apply to
subsections headers.