Prevent sidebar button scrolling out of view
When the sidebar was open and the main content was scrolled, the button was
scrolling out of view. Removing 'position: absolute' seems to avoid this;
positioning via the (already-present) 'fixed' property value seems to work.

The 'translateX' property is also removed, as this is already set via the
main '.sidebar-button' class selector block.