Move to esbuild and update npm packages
I didn't get anywhere with the two available esbuild Handlebars loaders,
so the assets build script instead precompiles the Handlebars templates.
https://handlebarsjs.com/installation/precompilation.html

The HTML formatter's digest function has been updated to produce 8-
character uppercase hashes in order to match those generated by esbuild.

The ESLint config was re-initiated and now uses its default parser
(Espree), so Babel is no longer required.

This commit does not update the formatters directory. When the
formatters directory is updated with the results of the new build
process, the 'formatters/html/fonts' directory should be deleted as the
font file is now placed in the dist directory along with the JS and CSS.

For the Elixir-based tests to result in up-to-date build dist contents I
found I had to empty the _build directory.