Improve consistency of type rendering weight across OSes
Fixes #1675

As the sidebar type is thinned in Mac OS via the alternative anti-aliasing, its weight can be increased from light to normal to avoid overly thin type in Linux and Windows.

The thin weight is preserved/used in places where differentiation from normal/bold is warranted.

Additionally, the normal/400 weight font was not being loaded, which resulted in type set to normal being rendered light. We're now loading light/300, normal/400 and bold/700 for the Lato sans serif font.