Reinstate heavier font-weight for sidebar list items for non-Apple
8e86d630 (Layout on apple-os and sidebar, 2023-03-29) resulted in
subitem list items in the sidebar having 300 weight on non-Apple, I
believe unintentionally. This reinstates their 400 weight for non-Apple.
(Although this means top-level and second-level items are the same
weight for non-Apple, I believe it was agreed the more readable font
weight took priority.)

To avoid this second declaration of font-weight 400 for non-Apple, we
could instead use !important for the first declaration, but this may
result in specificity issues if the design were to include different
weights in future.