Move autolink function parsing to language specific part
This makes it so that we do not have to deal with parsing
Erlang using the Elixir parser, things that are keywords in Elixir
are not in Erlang and vice versa.

The most recent example I came across was `c:do/1` which is valid
in Erlang but not in Elixir.