Fix so that source file and doc file are used correctly
With includes and generated files such as erl_parse.yrl
which source file and line a certain function/type/callback
is located in becomes quite complex. The documentation
can be located in yet another location, further complicating
things.

In addition, the relative directory from where we want the
source and doc information to be displayed can be different
as source information is used by create source_url links and
doc information is used to generate warnings.

So, this commit cleans things up and clearly separates source
and doc information.