Warn if there is a newer version available
Check hex.pm to see if there is a newer published version of ExDoc available
and warn if that's the case.

The GET request was a timeout of 1s and will not fail in case the response from
hex.pm is different from 200.

Jason isn't use only for testing anymore and HTTPoison was added as a
dependency as well.