Fix search bar input on focus 1px shift
Previously the `.search-input` set of styles defined `border: none`,
while its `:focus` state has 1px border set. This led to the
"micro-jump" of the placeholder text when the input field gets user
focus (and it "jumps" back when it loses one).

With this simple fix we always have invisible 1px border set to the
input, so it prevents "jumping".