Customize `search_data` for an extra (#2017)
This allows configuring the "search data" for any given extra. When you
do this, you take over the way that the extra appears both in
autocomplete and in search_data. The current system assumes that you
give this information in markdown, which I think is a safe assumption,
but we could do something like allow for an html_body or markdown_body
to be given, etc.

This will be used in Spark where we generate markdown files for DSL
documentation. With this change, we can add a function called
Spark.Dsl.search_data_for(Your.Dsl) which will improve the autocomplete
experience and the search experience in various ways (by adding more
things to autocomplete than just headers, and by customizing the type of
the search data nodes