Document the requirement to push releases
Currently there isn't any documentation about the requirement to push
releases/tags to the `source_url` to get links to specific versions of
the code. So this PR adds this documentation.

I couldn't find a general spot where `source_url` is talked about so I
added an admonition near where it is first introduced.